+++ Modules/m_gtdemo.cpp
@@ -31,7 +31,7 @@ TGtDemo* TGtDemo::pm = nullptr;
 
 bool TGtDemo::check_RT( int nrt )
 {
-    return ( nrt >RT_ICOMP && nrt < aMod.size());
+    return ( nrt >RT_ICOMP && nrt < static_cast<int>(aMod.size()));
 }
 ///
